test(ui): split follow-up steering e2e tests into their own file
chat-flow.follow-ups.e2e.test.ts crossed the 1000-line max-lines cap at
1079 lines after #124905, breaking check-lint-core on main. The seven
steer-behavior tests move to chat-flow.follow-up-steering.e2e.test.ts;
the shared expectChatBubbleAbove helper moves to chat-flow.test-support
(both files use it). No test bodies changed; no suppression added.
